+++ examples/landmark_uncertainty/main.py
@@ -26,8 +26,6 @@ from kale.embed.uncertainty_fitting import fit_and_predict
 from kale.interpret.uncertainty_quantiles import generate_fig_comparing_bins, generate_fig_individual_bin_comparison
 from kale.utils.download import download_file_by_url
 
-warnings.filterwarnings("error")
-
 
 def arg_parse():
     """Parsing arguments"""
@@ -124,10 +122,10 @@ def main():
                 for landmark in landmarks:
                     # Define Paths for this loop
                     landmark_results_path_val = os.path.join(
-                        cfg.DATASET.ROOT, base_dir, model, dataset, uncertainty_pairs_val + "_l" + str(landmark)
+                        cfg.DATASET.ROOT, base_dir, model, dataset, uncertainty_pairs_val + "_t" + str(landmark)
                     )
                     landmark_results_path_test = os.path.join(
-                        cfg.DATASET.ROOT, base_dir, model, dataset, uncertainty_pairs_test + "_l" + str(landmark)
+                        cfg.DATASET.ROOT, base_dir, model, dataset, uncertainty_pairs_test + "_t" + str(landmark)
                     )
 
                     fitted_save_at = os.path.join(save_folder, "fitted_quantile_binning", model, dataset)
